Add the concept of tasks to execute
When running the CLI, you can specify --task <task>,  which will make
the agent execute a given task. By default we execute the simple task,
which calls the LLM then terminates.

Add the ability to query multiple personas, summarize the results, or
simply summarize a query.

Indicate the task being processed when storing an interaction in Notion.

Update profiler to measure the before/after duration of a middleware.