Introduce Request and Response graphs
Request/Response graphs allows us to define a graph of requests which
will be executed when their dependencies are resolved. The response to
the request becomes available to the following node in the graph,
allowing it to make use of its dependencies response to complete a
given task.

For example, in this change we update the implementation of the
SummarizeMultiplePersonas task as we now can iterate over the list of
responses we want to summarize instead of having to pass around a number
representing the number of responses to use from the session history.

Another big improvement that comes with this change is treaded execution
which allows us to execute multiple calls to backends at the same time
instead of doing them sequentially, speeding up considerably the pace
at which the agent can go through a request graph, which used to be the
equivalent of a request graph where each request would have been done
sequentially.