Initial HTTP server + API
Naive implementation of an agent registry, only kept in memory.
When agents start, they will register with the agent registry.
It is possible to call the /agent/query in order to call the underlying
agent, allowing external users and agents to call the agent.
Each agent has its own agent registry locally that can also be queried
by others.